+++ Elements/src/Geometry/Line.cs
@@ -482,12 +482,13 @@ namespace Elements.Geometry
         /// Test if a point lies within tolerance of this line segment.
         /// </summary>
         /// <param name="point">The point to test.</param>
-        /// <param name="includeEnds">Consider a point at the endpoint as on the line.
+        /// <param name="includeEnds">Consider a point at the endpoint as on the line.</param>
+        /// <param name="tolerance">An optional distance tolerance.</param>
         /// When true, any point within tolerance of the end points will be considered on the line.
         /// When false, points precisely at the ends of the line will not be considered on the line.</param>
-        public bool PointOnLine(Vector3 point, bool includeEnds = false)
+        public bool PointOnLine(Vector3 point, bool includeEnds = false, double tolerance = Vector3.EPSILON)
         {
-            return Line.PointOnLine(point, Start, End, includeEnds);
+            return Line.PointOnLine(point, Start, End, includeEnds, tolerance);
         }
 
         /// <summary>
@@ -496,12 +497,13 @@ namespace Elements.Geometry
         /// <param name="point">The point to test.</param>
         /// <param name="start">The start point of the line segment.</param>
         /// <param name="end">The end point of the line segment.</param>
-        /// <param name="includeEnds">Consider a point at the endpoint as on the line.
+        /// <param name="includeEnds">Consider a point at the endpoint as on the line.</param>
+        /// <param name="tolerance">An optional distance tolerance.</param>
         /// When true, any point within tolerance of the end points will be considered on the line.
         /// When false, points precisely at the ends of the line will not be considered on the line.</param>
-        public static bool PointOnLine(Vector3 point, Vector3 start, Vector3 end, bool includeEnds = false)
+        public static bool PointOnLine(Vector3 point, Vector3 start, Vector3 end, bool includeEnds = false, double tolerance = Vector3.EPSILON)
         {
-            if (includeEnds && (point.IsAlmostEqualTo(start) || point.IsAlmostEqualTo(end)))
+            if (includeEnds && (point.IsAlmostEqualTo(start, tolerance) || point.IsAlmostEqualTo(end, tolerance)))
             {
                 return true;
             }
@@ -511,7 +513,7 @@ namespace Elements.Geometry
             if (lambda > 0 && lambda < 1)
             {
                 var pointOnLine = start + lambda * delta;
-                return pointOnLine.IsAlmostEqualTo(point);
+                return pointOnLine.IsAlmostEqualTo(point, tolerance);
             }
             return false;
         }